+++ src/components/HelloWorld.tsx
@@ -1,0 +1,5 @@
+import { Text } from 'react-native'
+
+export const HelloWorld = ({ text = 'Hello' }) => {
+  return <Text>{text}</Text>
+}